LOUD NOISES
The screenshot shares the impact point of the story. A cacophony of sounds play  out leading to this final point.

In order to get to this, I had to layer multiple different sounds on top of each other, and even apply effects such as distortion to get a reasonably high amount of gain.